+++ client/src/dispatcher/DispatcherView.tsx
@@ -1,5 +1,6 @@
-import { Box, Button, Center, Grid, Group, Stack, Title } from "@mantine/core"
+import { Center, Stack, Title } from "@mantine/core"
 import { useEffect } from "react"
+import Navbar from "../components/Navbar.js"
 import SharedTable from "../components/SharedTable"
 import store from "../store.js"
 import AssignButton from "./AssignButton.js"
@@ -9,8 +10,6 @@ import AssignButton from "./AssignButton.js"
 // maybe color the entire row based on a conditon.
 
 function DispatcherView() {
-  const setView = store((state) => state.setView)
-
   const setModalButtons = store((state) => state.setModalButtons)
   const setTableHeaders = store((state) => state.setTableHeaders)
 
@@ -21,47 +20,7 @@ function DispatcherView() {
 
   return (
     <>
-      <Box
-        w="100vw"
-        h={50}
-        mb={25}
-        // opacity = {.85}
-        sx={(theme) => ({
-          backgroundColor: "#488047", //: "#854040",
-          textAlign: "center",
-          padding: theme.spacing.md,
-          border: "",
-        })}
-      >
-        <Grid justify="center">
-          <Grid.Col span="auto">
-            <Button.Group>
-              <Button
-                variant="light"
-                color="gray"
-                onClick={() => {
-                  setView("default")
-                }}
-              >
-                <b>HOME</b>
-              </Button>
-              <Button
-                variant="light"
-                color="gray"
-                onClick={() => {
-                  setView("responder")
-                }}
-              >
-                RESPONDER
-              </Button>
-            </Button.Group>
-          </Grid.Col>
-          <Grid.Col span="auto"></Grid.Col>
-          <Grid.Col span={2} offset={2}>
-            <Group></Group>
-          </Grid.Col>
-        </Grid>
-      </Box>
+      <Navbar />
       <Center>
         <Stack>
           <Title order={1}>MEDEVAC Dispatch</Title>
