+++ client/src/responder/ResponderView.tsx
@@ -1,5 +1,6 @@
-import { Box, Button, Center, Grid, Group, Stack, Title } from "@mantine/core"
+import { Center, Stack, Title } from "@mantine/core"
 import { useEffect } from "react"
+import Navbar from "../components/Navbar.js"
 import SharedTable from "../components/SharedTable"
 import store from "../store.js"
 import CompleteButton from "./CompleteButton"
@@ -10,7 +11,6 @@ import RoleTwoButton from "./RoleTwoButton.js"
 // maybe color the entire row based on a conditon.
 
 function ResponderView() {
-  const setView = store((state) => state.setView)
   const setModalButtons = store((state) => state.setModalButtons)
   const setTableHeaders = store((state) => state.setTableHeaders)
 
@@ -20,49 +20,9 @@ function ResponderView() {
   }, [])
 
   return (
-    <Center>
-      <Stack>
-        <Box
-          w="100vw"
-          h={50}
-          mb={25}
-          // opacity = {.85}
-          sx={(theme) => ({
-            backgroundColor: "#488047", //: "#854040",
-            textAlign: "center",
-            padding: theme.spacing.md,
-            border: "",
-          })}
-        >
-          <Grid justify="center">
-            <Grid.Col span="auto">
-              <Button.Group>
-                <Button
-                  variant="light"
-                  color="gray"
-                  onClick={() => {
-                    setView("default")
-                  }}
-                >
-                  <b>HOME</b>
-                </Button>
-                <Button
-                  variant="light"
-                  color="gray"
-                  onClick={() => {
-                    setView("dispatcher")
-                  }}
-                >
-                  DISPATCHER
-                </Button>
-              </Button.Group>
-            </Grid.Col>
-            <Grid.Col span="auto"></Grid.Col>
-            <Grid.Col span={2} offset={2}>
-              <Group></Group>
-            </Grid.Col>
-          </Grid>
-        </Box>
+    <>
+      <Navbar></Navbar>
+      <Center>
         <Stack>
           <Title order={1}>MEDEVAC Assignment</Title>
           <Title order={5}>SE Texas</Title>
